tests/wpt/tests/css/css-text/parsing/text-wrap-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing text-wrap with invalid values</title>
<link rel="author" title="Tim Nguyen" href="https://github.com/nt1m">
<link rel="help" href="https://drafts.csswg.org/css-text-4/#text-wrap">
<meta name="assert" content="text-wrap supports only the grammar '<text-wrap-mode> || <text-wrap-style>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("text-wrap", "normal");
test_invalid_value("text-wrap", "none");
test_invalid_value("text-wrap", "wrap nowrap");
test_invalid_value("text-wrap", "pretty balance");
test_invalid_value("text-wrap", "balance stable");
test_invalid_value("text-wrap", "stable pretty");
test_invalid_value("text-wrap", "delicious wrap");
test_invalid_value("text-wrap", "5px");
</script>
</body>
</html>
